# Overlap qk norm with two streams